Fixed: long indexes on mem-map bounds.

+++ AbstractRendering/src/ar/glyphsets/MemMapList.java
@@ -135,7 +135,7 @@ public class MemMapList implements GlyphSet, GlyphSet.RandomAccess, Iterable<Gly
 			double minX=Double.MAX_VALUE, minY=Double.MAX_VALUE, maxX=Double.MIN_VALUE, maxY=Double.MIN_VALUE;
 			BigFileByteBuffer buffer = this.buffer.get();
 
-			for (int i=0; i<size();i++) {
+			for (long i=0; i<size();i++) {
 				long recordOffset = (i*recordSize)+headerOffset;
 				buffer.position(recordOffset);
 				double x = value(buffer, 0);
